perf: pack Metal jump indices to uint8

+++ macos/MetalFieldKernels.h
@@ -583,7 +583,7 @@ kernel void jacobian_affine_walk_jump_table(constant ulong* p_xyz [[buffer(0)]],
                                             device uint* out_infinity [[buffer(4)]],
                                             constant uint& count [[buffer(5)]],
                                             constant uint& steps [[buffer(6)]],
-                                            constant uint* jump_indices [[buffer(7)]],
+                                            constant uchar* jump_indices [[buffer(7)]],
                                             constant ulong* jump_distances [[buffer(8)]],
                                             device ulong* out_distances [[buffer(9)]],
                                             constant ulong& dp_mask [[buffer(10)]],
@@ -600,7 +600,7 @@ kernel void jacobian_affine_walk_jump_table(constant ulong* p_xyz [[buffer(0)]],
   ulong distance = 0;
 
   for (uint step = 0; step < steps; step++) {
-    uint jump_index = jump_indices[jump_base + step];
+    uint jump_index = (uint)jump_indices[jump_base + step];
     distance += jump_distances[jump_index];
     uint q_base = jump_index << 3;
     JacobianValue out = jacobian_add_affine_values(x0, x1, x2, x3, y0, y1, y2, y3, z0, z1, z2, z3, inf,
@@ -625,7 +625,7 @@ kernel void jacobian_affine_walk_jump_table_steps8(constant ulong* p_xyz [[buffe
                                                    device uint* out_infinity [[buffer(4)]],
                                                    constant uint& count [[buffer(5)]],
                                                    constant uint& steps [[buffer(6)]],
-                                                   constant uint* jump_indices [[buffer(7)]],
+                                                   constant uchar* jump_indices [[buffer(7)]],
                                                    constant ulong* jump_distances [[buffer(8)]],
                                                    device ulong* out_distances [[buffer(9)]],
                                                    constant ulong& dp_mask [[buffer(10)]],
@@ -643,7 +643,7 @@ kernel void jacobian_affine_walk_jump_table_steps8(constant ulong* p_xyz [[buffe
   ulong distance = 0;
 
   for (uint step = 0; step < 8; step++) {
-    uint jump_index = jump_indices[jump_base + step];
+    uint jump_index = (uint)jump_indices[jump_base + step];
     distance += jump_distances[jump_index];
     uint q_base = jump_index << 3;
     JacobianValue out = jacobian_add_affine_values(x0, x1, x2, x3, y0, y1, y2, y3, z0, z1, z2, z3, inf,
